requirements.txt
llama-index-core==0.12.24.post1
llama-index-embeddings-huggingface==0.5.2
llama-index-vector-stores-chroma==0.4.1
transformers==4.44.1
chromadb==0.6.3